converters/waveformer/requirements.txt
fastapi[standard]
faststream[nats]
httpx

conductor-python==1.1.10
boto3

pedalboard